+++ animation_and_polting.py
@@ -1,0 +1,20 @@
+import matplotlib.pyplot as plt
+import time
+from matplotlib.animation import FuncAnimation
+
+
+# Tf: float = 42.0):
+def plot(self, parameter_list):
+     pass
+
+def __update(self, frame):
+     u =_fitzhugh_nagumo_initial_conditions()
+    for i in range(10):
+        u_new = u + k*( eps*(L@u + bc) + (u**2 - u**3 - a*u + a*u**2) )
+        u[:] = u_new
+
+    ln.set_data(x, u)
+    ax.set_title('t = {}'.format(10*frame*k))
+    
+def animation(self):
+pass
